Builds an object that will be responsible for representing what kind of motion we want to display.
This commit also changes the activity's name and implements the list in the ResponsiveListFragment. This should be split in another commit but sorry. Another point to say is that we will only have two activities in this project. One will be responsible for handling the navigation fragments and the other one will be responsible for handling the animation fragments.